% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnageneticMatrix.R
\name{AnageneticMatrix}
\alias{AnageneticMatrix}
\title{make the anagenetic rate matrix function}
\usage{
AnageneticMatrix(num_states, rate = NULL)
}
\description{
make the anagenetic rate matrix function
}
