% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{approximatorVersion}
\alias{approximatorVersion}
\title{Approximation algorithms}
\description{
A list of approximators for tensorphylo
}
\details{
Use these variables to specify the type of approximator to use.
There are five options:\tabular{lcl}{
   \strong{Setting} \tab \strong{Index} \tab \strong{Interpretation} \cr
   \code{approximatorVersion$AUTO_TUNING} \emph{(default)} \tab 0 \tab Use the approximator that works the best after the first few calculations. \cr
   \code{approximatorVersion$SEQUENTIAL_OPTIMIZED} \tab 1 \tab Integrate each branch one at a time, but share extinction probability calculations. \cr
   \code{approximatorVersion$SEQUENTIAL_BRANCHWISE} \tab 2 \tab Integrate each branch one at a time, with separate extinction probability calculations. \cr
   \code{approximatorVersion$PARALLEL_OPTIMIZED} \tab 3 \tab Parallelize integration over branches, but share extinction probability calclations. \cr
   \code{approximatorVersion$PARALLEL_BRANCHWISE} \tab 4 \tab Parallelize integration over branches, with separate extinction probability calculations. \cr
}
}
\examples{
# create an empty tensorphylo object
tp <- new(TensorPhylo, 4)

# specify a debug setting
tp$setLikelihoodApproximator( approximatorVersion$AUTO_TUNING )

# this is equivalent to:
tp$setLikelihoodApproximator( 0 )
}
