% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CladoEvents.R
\name{CladoEvents}
\alias{CladoEvents}
\title{Cladogenetic Event array}
\description{
The \code{CladoEvents} class.
}
\details{
This class is for sparse arrays of cladogenetic events.
The \emph{i,j,k}'th element containts the probability that an ancestral lineage in state \emph{i} leaves a left daughter in state \emph{j} and a right daughter in state \emph{k} during a speciation event.

Objects of this class have only need to have a subset of possible events specified; unspecified events are assumed to occur with probability 0.
Because the elements of this array represent probabilities, the sum of all probabilities for a given ancestral state must be 1.
The class enforces that all values are between 0 and 1, and that ancestral probabilities sum to 1.

For models with time-heterogeneous cladogenetic events, use the generic \code{c.CladoEvents} to create a vector of event array (of class \code{CladoEventsList}):
\preformatted{
O_1 <- makeCladogeneticEvents(3)
O_2 <- makeCladogeneticEvents(3)
Os  <- c(O_1, O_2)
}
\subsection{Constructors}{

Create a new object of class \code{CladoEvents}.
\itemize{
\item \code{makeCladogeneticEvents(dim)} where \code{dim} is an integer. Creates an empty array.
\item \code{new(CladoEvents, dim)}. Equivalent to \code{makeCladogeneticEvents(dim)}.
}
}

\subsection{Accessors}{

\code{CladoEvents} provides intuitive access to elements of the array, but enforces some behavior.
\itemize{
\item \verb{[i,j,k]} returns value corresponding to an ancestor in state \emph{i} leaving left and right descendants in states \emph{j} and \emph{k}, respectively.
\item \verb{[i,j,k] <- y} sets the value of the _ijk_th element to y. \strong{\emph{Will result in an error if you attempt to set a non-event (i = j = k), or attempt to set an element to a value outside of \link{0,1} bounds.}}
}
}
}
\examples{
# create a cladogenetic event array.
O <- makeCladogeneticEvents(3)

# set the probability of the ancestor in state 1 leaving left daughter in state 1 and right daughter in state 2
O[1,1,2] <- 0.2
\dontrun{
# setting a diagonal value is prohibited
O[1,1,1] <- 0.1

# likewise, setting a value outside of [0,1] is prohibited.
O[1,1,2] <- -0.1
O[1,1,2] <-  1.1
}

# construct a time-homogeneous cladogenetic model
O  <- makeCladogeneticEvents(4)
tp <- new(TensorPhylo, 4)
tp$setOmegaConstant(O)

# construct a time-heterogeneous cladogenetic model
O_1 <- makeCladogeneticEvents(4)
O_2 <- makeCladogeneticEvents(4)
Os  <- c(O_1, O_2)
t   <- 0.1
tp$setOmegaTimeVarying(t, Os)
}
