% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{conditionalProbability}
\alias{conditionalProbability}
\title{Conditional probabilities}
\description{
A list of conditional probability options for tensorphylo
}
\details{
These variables to specify the type of conditional probability calculations.
There are nine options:\tabular{lcl}{
   \strong{Setting} \tab \strong{Index} \tab \strong{Interpretation} \cr
   \code{conditionalProbability$TIME} \emph{(default)} \tab 0 \tab Don't condition on anything. \cr
   \code{conditionalProbability$ROOT_SURVIVAL} \tab 1 \tab Assume that the two lineages at the root each had to leave at least one extant sample. \cr
   \code{conditionalProbability$ROOT_MRCA} \tab 2 \tab Assume that the two lineages at the root each had to leave at least one extant sample, \emph{AND} that there was a speciation event at the root. \strong{This is equivalent to conditioning on survival in most other programs.} \cr
   \code{conditionalProbability$ROOT_SAMPLING} \tab 3 \tab Assume that the two lineages at the root each had to leave at least one sample (extant or extinct). \cr
   \code{conditionalProbability$ROOT_SAMPLING_AND_MRCA} \tab 8 \tab Assume that the two lineages at the root each had to leave at least one sample (extant or extinct), \emph{AND} that there was a speciation event at the root. \cr
   \code{conditionalProbability$STEM_SURVIVAL} \tab 4 \tab Assume that the one lineage at the stem had to leave at least one extant sample This is equivalent to conditioning on survival (with a stem) in most programs. \cr
   \code{conditionalProbability$STEM_ONE_SAMPLE} \tab 5 \tab Assume that the one lineage at the stem had to leave at least one sample (extant or extinct). \cr
   \code{conditionalProbability$STEM_TWO_EXT_SAMPLES} \tab 6 \tab Assume that the one lineage at the stem had to leave at least two extant samples. \cr
   \code{conditionalProbability$STEM_TWO_SAMPLES} \tab 7 \tab Assume that the one lineage at the stem had to leave at least two samples (extant or extinct). \cr
}
}
\examples{
# create an empty TensorPhyloInstance object
tp <- new(TensorPhyloInstance, 4)

# condition on survival of the root lineages
tp$setConditionalProbabilityType( conditionalProbability$ROOT_SURVIVAL )

# this is equivalent to:
tp$setConditionalProbabilityType( 1 )
}
