% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{debugMode}
\alias{debugMode}
\title{tensorphylo debug modes}
\description{
A list of debug modes for tensorphylo
}
\details{
Use these variables to specify debug modes
There are five options:\tabular{lcl}{
   \strong{Setting} \tab \strong{Index} \tab \strong{Interpretation} \cr
   \code{debugMode$DBG_NONE} \emph{(default)} \tab 0 \tab No debug info \cr
   \code{debugMode$DBG_PRINT} \tab 1 \tab Debug info printed to screen \cr
   \code{debugMode$DBG_FILE} \tab 2 \tab Debug info printed to file \cr
}
}
\examples{
# create an empty tensorphylo object
tp <- new(TensorPhylo, 4)

# specify a debug setting
tp$debugMode( debugMode$DBG_PRINT )

# this is equivalent to:
tp$debugMode( 0 )
}
