% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{integrationScheme}
\alias{integrationScheme}
\title{Integration algorithms}
\description{
A list of numerical integration schemes for tensorphylo
}
\details{
These variables to specify the type of numerical integration algorithm to use.
There are four options:\tabular{lcl}{
   \strong{Setting} \tab \strong{Index} \tab \strong{Interpretation} \cr
   \code{integrationScheme$EULER} \tab 0 \tab The explicit Euler integrator. \cr
   \code{integrationScheme$RUNGE_KUTTA4} \tab 1 \tab The Runge-Kutta 4 integrator. \cr
   \code{integrationScheme$RUNGE_KUTTA54} \emph{(default)} \tab 2 \tab The Runge-Kutta 4(5) integrator (AKA the Runge-Kutta-Fehlberg algorithm). \cr
   \code{integrationScheme$RUNGE_KUTTA_DOPRI5} \tab 3 \tab The Dormand–Prince integrator. \cr
}
}
\examples{
# create an empty TensorPhyloInstance object
tp <- new(TensorPhyloInstance, 4)

# specify a debug setting
tp$setIntegrationScheme( integrationScheme$RUNGE_KUTTA_DOPRI5 )

# this is equivalent to:
tp$setIntegrationScheme( 3 )
}
