% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CladoEvents.R
\name{makeCladogeneticEvents}
\alias{makeCladogeneticEvents}
\title{Cladogenetic array constructor.}
\usage{
makeCladogeneticEvents(num_states)
}
\arguments{
\item{num_states}{The number of states.}
}
\value{
A default object of class \code{\link{CladoEvents}}.
}
\description{
Creates a cladogenetic event  array
}
\details{
Creates an default object of class \code{\link{CladoEvents}}.
}
\examples{
# create a cladogenetic event array.
O <- makeCladogeneticEvents(3)

# set the probability of the ancestor in state 1 leaving left daughter in state 1 and right daughter in state 2
O[1,1,2] <- 0.2

}
\seealso{
\code{\link{CladoEvents}}
}
