% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{makeProbabilityMatrix}
\alias{makeProbabilityMatrix}
\title{Transition matrix constructor.}
\usage{
makeProbabilityMatrix(num_states)
}
\arguments{
\item{num_states}{The number of states.}
}
\value{
A num_states x num_states matrix of class \code{\link{ProbabilityMatrix}}.
}
\description{
Creates a transition-probability matrix.
}
\details{
Creates a num_states by num_states object of class \code{\link{ProbabilityMatrix}}.
}
\examples{
# create a 4x4 probability matrix.
P <- makeProbabilityMatrix(4)

# set the rate from 1 to 2 to 0.2
P[1,2] <- 0.2

}
\seealso{
\code{\link{ProbabilityMatrix}}
}
