% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateMatrix.R
\name{makeRateMatrix}
\alias{makeRateMatrix}
\title{Rate Matrix constructor.}
\usage{
makeRateMatrix(num_states, rate = NULL)
}
\arguments{
\item{num_states}{The number of states.}

\item{rate}{(optional) If specified, creates an equal-rates matrix with average rate \code{rate}.}
}
\value{
A num_states x num_states matrix of class \code{\link{RateMatrix}}.
}
\description{
Creates a rate matrix.
}
\details{
Creates a num_states by num_states object of class \code{\link{RateMatrix}}.
}
\examples{
# create a 4x4 rate matrix with average rate of 0.1 (ie a Jukes-Cantor model with mu = 0.1).
Q <- makeRateMatrix(4, 0.1)

# set the rate from 1 to 2 to 0.2
Q[1,2] <- 0.2

}
\seealso{
\code{\link{RateMatrix}}
}
