% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{ProbabilityMatrix}
\alias{ProbabilityMatrix}
\title{Probability Matrix}
\description{
The \code{ProbabilityMatrix} class.
}
\details{
This class is for transition-probability matrices, which describe the probability of changing from state \emph{i} (rows) to state \emph{j} (columns).
These matrices are useful for "synchronous" state-change events (where all lineages change simultaneously) at some time \emph{t}.

Objects of the class \code{ProbabilityMatrix} have elements that range from 0 to 1 such that the sum of each row is one.
Attempting to set the value of a diagonal element is prohibited, as they are computed automatically from the off-diagonal elements.
Likewise, the class enforces that element values \code{x} must be \verb{0 <= x <= 1}.

For models with time-heterogeneous rate matrices, use the generic \code{ProbabilityMatrix} to create a vector of rate matrices (of class \code{ProbabilityMatrixList}):
\preformatted{
P_1 <- makeProbabilityMatrix(3)
P_2 <- makeProbabilityMatrix(3)
Ps  <- c(P_1, P_2)
}
\subsection{Constructors}{

Create a new object of class \code{ProbabilityMatrix}.
\itemize{
\item \code{makeProbabilityMatrix(dim)} where \code{dim} is an integer. Creates a \code{dim} x \code{dim} probability matrix with 1s along the diagonal.
\item \code{new(ProbabilityMatrix, dim)}. Equivalent to \code{makeProbabilityMatrix(dim)}.
}
}

\subsection{Methods}{

Call a method on a \code{makeProbabilityMatrix} object \code{P} with \code{P$methodName(arguments)}.
\itemize{
\item \code{getMatrix()}: Returns a standard R \link[base:matrix]{base::matrix}.
}
}

\subsection{Accessors}{

\code{ProbabilityMatrix} provides intuitive access to elements of the rate matrix, but enforces some behavior.
\itemize{
\item \verb{[i,j]} returns value in the _i_th row and _j_th column.
\item \verb{[i,j] <- y} sets the value of the _ij_th element to y. \strong{\emph{Will result in an error if you attempt to set a diagonal value, or attempt to set an off-diagonal element to below 0.0 or above 1.0.}}
}
}
}
\examples{
# create a 4x4 probability matrix
P <- makeProbabilityMatrix(4)

# set the rate from 1 to 2 to 0.2
P[1,2] <- 0.2

\dontrun{
# setting a diagonal value is prohibited
P[1,1] <- 0.1

# likewise, setting a value outside of [0,1] is prohibited.
P[1,2] <- -0.1
P[1,2] <- 1.1
}

# construct a state-change model with one event at time t
# (we have to use a vector of P because there may be more than one
# time point, i.e., length(t) > 1).
P  <- makeProbabilityMatrix(4)
P[1,2] <- 0.1
t <- 0.1
tp <- new(TensorPhyloInstance, 4)
tp$setZeta(t, c(P))

}
