% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readNexusData}
\alias{readNexusData}
\title{Read Nexus formatted data}
\arguments{
\item{file}{path to a nexus file with one DATA block, that contains a single discrete character.}
}
\value{
A matrix formatted for input to tensorphylo. Named rows correspond to species, columns correspond to character states. A value of 1 indicates that the character is consistent with the observed data.
}
\description{
Read data in the Nexus file format for tensorphylo.
In principle, works with any nexus datatype and symbols (I haven't done extensive testing), and handles ambiguity/missing data.
}
