% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tensoRphylo}
\alias{tensoRphylo}
\title{tensoRphylo}
\description{
An R Interface for the TensorPhylo Library
}
\details{
An R interface to tensorphylo, a high-performance library for probability calculations under a general class of state-dependent birth-death models.
}
\author{
Michael R May
}
