% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorPhyloInstance.R
\name{TensorPhyloInstance}
\alias{TensorPhyloInstance}
\alias{makeTensorPhylo}
\title{API documentation}
\description{
API documentation
}
\section{Constructor}{

}

\section{Models and Parameters}{

\subsection{Lineage-specific diversification events}{
\subsection{Speciation events}{
}

\subsection{Extinction events}{
}

\subsection{Sampling events}{
}

\subsection{Destructive-sampling events}{
}

}

\subsection{Mass-diversification events}{
\subsection{Mass-speciation events}{
}

\subsection{Mass-extinction events}{
}

\subsection{Mass-sampling events}{
}

\subsection{Mass-destructive-sampling events}{
}

}

\subsection{State-change events}{
\subsection{Anagenetic events}{
}

\subsection{Cladogenetic events}{
}

}

\subsection{Root Frequencies}{
}

\subsection{Conditional probabilities}{
}
}

\section{Numerical settings}{

}

