% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{make the transition probability matrix function}
\usage{
TransitionMatrix(num_states)
}
\description{
make the transition probability matrix function
}
