% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorPhyloInstance.R
\name{TensorPhyloInstance}
\alias{TensorPhyloInstance}
\alias{makeTensorPhylo}
\alias{TensorPhylo}
\title{Instance}
\usage{
makeTensorPhylo(tree, data = NULL, nstates = NULL)
}
\description{
The tensorphylo interface.
}
\details{
This is the general interface for working with the tensorphylo high-performance library for state-dependent birth-death models.
}
\section{Creating TensorPhyloInstance objects}{

\subsection{Hmm}{
}
}

\section{Speciation rates}{
NA
}

\section{Extinction rates}{
NA
}

\section{Sampling rates}{
NA
}

\section{Destructive-sampling rates}{
NA
}

\section{Mass-speciation events}{
NA
}

\section{Mass-extinction events}{
NA
}

\section{Mass-sampling events}{
NA
}

\section{Mass-destructive-sampling events}{
NA
}

\section{Anagenetic state-change events}{
NA
}

\section{Cladogenetic state-change events}{
NA
}

\section{Mass-anagenetic-state-change events}{
NA
}

